% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_fastq.R
\name{count_GC}
\alias{count_GC}
\title{Number of G and C}
\usage{
count_GC(secuencia)
}
\arguments{
\item{secuencia}{Input DNA sequence}
}
\value{
GC percentage value
}
\description{
This function calculates the percentage of GC in a given sequence
}
