% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_fastq.R
\name{cuartil}
\alias{cuartil}
\title{First quartile for Quality scores dataframe}
\usage{
cuartil(quality_df)
}
\arguments{
\item{quality_df}{Dataframe with all the quality scores; positions as the column names, values in names}
}
\value{
A vector with all the first quartile values for the scores in each position
}
\description{
This function calculates the first quartile of the quality scores in each position of the readsGC content for a set of sequences in a vector
}
