% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_fastq.R
\name{GC_content}
\alias{GC_content}
\title{Calculation of GC content for a set of sequences}
\usage{
GC_content(sequences_vector)
}
\arguments{
\item{sequences_vector}{Vector with all the sequences in the FASTQ file}
}
\value{
A vector with all the GC percentage values ready for plotting
}
\description{
This function calculates GC content for a set of sequences in a vector
}
