% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_fastq.R
\name{read_fastq}
\alias{read_fastq}
\title{Read a FASTQ file}
\usage{
read_fastq(file)
}
\arguments{
\item{file}{Input FASTQ file}
}
\value{
A dataframe with 2 columns: sequences and scores
}
\description{
This function reads a FASTQQ file and returns the sequences in it and their ASCII scores in a dataframe
}
