% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_fastq.R
\name{reads_length}
\alias{reads_length}
\title{Length of reads}
\usage{
reads_length(sequences_vector)
}
\arguments{
\item{sequences_vector}{Vector with all the sequences in the FASTQ file}
}
\value{
Prints a summary of the values and returns the mean
}
\description{
This function calculates the length of the reads in the FASTQ file
}
