% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_fastq.R
\name{scoring}
\alias{scoring}
\title{Quality Scores from ASCII values}
\usage{
scoring(ascii_vector, read_length = 100)
}
\arguments{
\item{read_length}{Length of reads (100 by default)}

\item{sequences_vector}{Vector with all the ASCII scores for the sequences in the FASTQ file}
}
\value{
A dataframe with all the quality scores for each position
}
\description{
This function calculates quality scores given the sequences with ASCII characters
}
